% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abundance.R
\docType{data}
\name{Abundance}
\alias{Abundance}
\title{Abundance data sample}
\format{
The \code{Abundance} is a matrix covering two stations and three species.
}
\source{
Karasiewicz S., and Lefebvre A. (2022). Environmental Impact on Harmful Species Pseudo-nitzschia spp. and Phaeocystis globosa Phenology and Niche. \emph{JMSE} 10(2), 174. \doi{10.3390/jmse10020174}.
}
\usage{
data(Abundance)
}
\description{
The \code{Abundance} data are a sample of the entire data set used in Karasiewicz et al.(2022).
}
\references{
Karasiewicz S., and Lefebvre A. (2022). Environmental Impact on Harmful Species Pseudo-nitzschia spp. and Phaeocystis globosa Phenology and Niche. \emph{JMSE} 10(2), 174. \doi{10.3390/jmse10020174}.
}
\keyword{datasets}
